Interfaces para Android

+++ Practica1/Engine/src/main/java/com/example/engine/IFont.java
@@ -1,7 +1,7 @@
 package com.example.engine;
 
 public interface IFont {
-    public int getSize();
+    public float getSize();
     public boolean isBold();
     public void setSize(float size);
 }
